refactor : IDE-121 타입 체크 및 필요 없는 파일 삭제
