+++ src/app/ide/FileTree/Node.tsx
@@ -7,7 +7,8 @@ import { NodeRendererProps } from 'react-arborist';
 import { FileDiv, IsDirty, IsNotDirty, NodeContainer } from './FileTree.styles';
 import { useFileTreeStore } from '@/store/useFileTreeStore';
 import { FileNodeType } from '@/types/IDE/FileTree/FileDataTypes';
-import { findLanguage, isCorrectName } from '@/utils/fileTreeUtils';
+import { findLanguage } from '@/utils/filetree/findFileLangUtils';
+import { isCorrectName } from '@/utils/filetree/fileTreeUtils';
 import useHandleOpenFile from '@/hooks/useHandleOpenFile';
 import LanguageIcon from './LanguageIcon';
 import useHandleCreateFile from '@/hooks/useHandleCreateFile';
